Update teams.inc.php
I was sorting the Teams and 1.10.4 and noticed clicking the Status header in order to sort the Teams by 'active' and 'disabled' actually made all the teams disappear.  Problem was the sortOptions assigned the 'status' header to a non-existing field 'isenabled'.  In the ost_teams table the field is called 'flags'.  I changed line 7 here to make it work on my installation.

Does this make sense?  I don't see any other changes in develop that conflict.